mcp: carve the HTTP gate (pure move)
